requirements.txt
asgiref==3.8.1
Django==4.2
django-seed==0.3.1
djangorestframework==3.15.1
Faker==24.14.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.11.0
